fix(filter): simplify component
