+++ src/components/filter/__tests__/Filter.jest.spec.js
@@ -2,18 +2,16 @@ import { mount } from '@vue/test-utils';
 import axe from '@/../test/unit/helpers/axe-helper';
 import VsFilter from '../VsFilter.vue';
 
-const applyButtonText = 'Apply';
-const filterButtonText = 'Filter';
-const filterId = 'modal-1';
+const filterId = 'filter-1';
+const filterLabel = 'filter';
 const defaultSlotContent = 'Slot content';
 let wrapper;
 
 beforeEach(() => {
     wrapper = mount(VsFilter, {
         propsData: {
-            applyButtonText,
-            filterButtonText,
             filterId,
+            filterLabel,
         },
         slots: {
             default: defaultSlotContent,
@@ -27,39 +25,30 @@ describe('VsFilter', () => {
     });
 
     describe(':props', () => {
-        it('filterButtonText - should render a button with the value passed', () => {
-            const filterButton = wrapper.find('.vs-filter__toggle');
-
-            expect(filterButton.text()).toContain(filterButtonText);
+        it('filterId - should render the component with the ID value passed', () => {
+            expect(wrapper.find(`#${filterId}`).exists()).toBe(true);
         });
 
-        it('applyButtonText - should render a button component within the modal with the passed value', async() => {
-            const modal = wrapper.findComponent({
-                name: 'VsModal',
-            });
+        it('filterLabel - should render the fieldset with a legend containing the value passed', () => {
+            const filterLegend = wrapper.find('.vs-filter__legend');
 
-            const applyButton = modal.findAllComponents({
-                name: 'VsButton',
-            })[1];
-
-            expect(applyButton.exists()).toBe(true);
-            expect(applyButton.classes()).toContain('vs-filter__apply');
-            expect(applyButton.text()).toBe(applyButtonText);
-        });
-
-        it('filterId - should render a modal with the `modal-id` prop with the value passed', () => {
-            const modal = wrapper.findComponent({
-                name: 'VsModal',
-            });
-
-            expect(modal.exists()).toBe(true);
-            expect(modal.props('modalId')).toBe(filterId);
+            expect(filterLegend.text()).toContain(filterLabel);
         });
     });
 
     describe(':slots', () => {
         it('renders content inserted into the default slot', () => {
             expect(wrapper.text()).toContain(defaultSlotContent);
+        });
+    });
+
+    describe(':emits', () => {
+        it(':filter-updated - should be emitted when a filter item is changed', async() => {
+            const filterPanel = wrapper.find('.vs-filter__panel');
+
+            await filterPanel.trigger('change');
+
+            expect(wrapper.emitted('filter-updated')).toBeTruthy();
         });
     });
