+++ src/components/filter/index.ts
@@ -1,9 +1,7 @@
 import VsFilter from './VsFilter.vue';
-import VsFilterPanel from './components/VsFilterPanel.vue';
 import VsFilterSection from './components/VsFilterSection.vue';
 
 export default {
     VsFilter,
-    VsFilterPanel,
     VsFilterSection,
 };
